requirements.txt
streamlit
groq
chromadb
pypdf
openpyxl
pillow